Source/requirements.txt
paramiko
pycrypto
psutil
mysql-connector-python